refactor: move shared product types in separate file
