+++ src/api/product/product.types.ts
@@ -1,29 +1,5 @@
-export enum Attributes {
-    family = 'family',
-    size = 'size',
-    careLevel = 'care-level',
-    toxic = 'toxic',
-    lightRequirements = 'light-requirements',
-    height = 'height',
-}
-
-export enum Sizes {
-    small = 'S',
-    medium = 'M',
-    large = 'L',
-}
-
-export enum Care {
-    easy = 'easy',
-    medium = 'medium',
-    hard = 'hard',
-}
-
-export enum Light {
-    low = 'low',
-    medium = 'medium',
-    high = 'high',
-}
+import type { Sizes, Light, Care } from '../../types/product.types';
+import type { Variant, CustomVariant } from '../../types/product.types';
 
 export interface ProductResponse {
     id: string;
@@ -44,21 +20,6 @@ export interface ProductResponse {
     };
 }
 
-export interface Variant {
-    id: number;
-    sku: string;
-    prices: Price[];
-    images: Image[];
-    attributes: (
-        | FamilyAttribute
-        | SizeAttribute
-        | CareAttribute
-        | ToxicAttribute
-        | LightAttribute
-        | HeightAttribute
-    )[];
-}
-
 export type AttributesList = (
     | [string, string]
     | [string, Sizes]
@@ -68,93 +29,12 @@ export type AttributesList = (
     | [string, number]
 )[];
 
-export interface Price {
-    id: string;
-    value: {
-        type: string;
-        currencyCode: string;
-        centAmount: number;
-        fractionDigits: number;
-    };
-    discounted?: {
-        value: {
-            type: string;
-            currencyCode: string;
-            centAmount: number;
-            fractionDigits: number;
-        };
-    };
-    key: string;
-}
-
-export interface Image {
-    url: string;
-    label: string;
-    dimensions: {
-        w: number;
-        h: number;
-    };
-}
-
-interface FamilyAttribute {
-    name: Attributes.family;
-    value: string;
-}
-
-interface SizeAttribute {
-    name: Attributes.size;
-    value: {
-        key: Sizes;
-        label: Sizes;
-    };
-}
-
-interface CareAttribute {
-    name: Attributes.careLevel;
-    value: {
-        key: Care;
-        label: Care;
-    };
-}
-
-interface ToxicAttribute {
-    name: Attributes.toxic;
-    value: boolean;
-}
-
-interface HeightAttribute {
-    name: Attributes.height;
-    value: number;
-}
-
-interface LightAttribute {
-    name: Attributes.lightRequirements;
-    value: {
-        key: Light;
-        label: Light;
-    };
-}
-
 export interface CustomProduct {
     id: string;
     name: string;
     description: string;
     masterVariant: CustomVariant;
     variants: CustomVariant[];
-}
-
-export interface CustomVariant {
-    id: number;
-    sku: string;
-    price: number;
-    images: Image[];
-    family: string;
-    size: Sizes;
-    care: Care;
-    toxic: boolean;
-    height: number;
-    light: Light;
-    discount: number | null;
 }
 
 export interface CustomAttributes {
