+++ src/components/blocks/ProductDetails/parts/SizeAttribute/SizeAttribute.test.tsx
@@ -1,5 +1,5 @@
 import { render, screen } from '@testing-library/react';
-import { Care, Light, Sizes } from '../../../../../api/product/product.types';
+import { Care, Light, Sizes } from '../../../../../types/product.types';
 import type { CustomProduct } from '../../../../../api/product/product.types';
 import { SizeAttribute } from './SizeAttribute';
 import '@testing-library/jest-dom';
