+++ src/components/blocks/ProductDetails/parts/SizeAttribute/SizeAttribute.types.ts
@@ -1,8 +1,6 @@
-import type {
-    CustomProduct,
-    CustomVariant,
-    Sizes,
-} from '../../../../../api/product/product.types';
+import type { CustomProduct } from '../../../../../api/product/product.types';
+
+import type { CustomVariant, Sizes } from '../../../../../types/product.types';
 
 export interface SizeAttributeProps {
     product: CustomProduct;
