+++ src/components/ui/ImagesSlider/ImagesSlider.types.ts
@@ -1,4 +1,4 @@
-import type { Image } from './../../../api/product/product.types';
+import type { Image } from '../../../types/product.types';
 
 export interface ImagesSliderProps {
     images: Image[];
